ml-model-api/requirements.txt
Flask==2.3.2
torch==1.13.1
Pillow==9.3.0
